+++ bika/lims/browser/js/bika.lims.worksheet.js
@@ -881,13 +881,17 @@
       el = event.currentTarget;
       $(el).prepOverlay({
         subtype: "ajax",
-        filter: "h1,span.remarks_history",
+        filter: "h1,div.remarks-widget",
         config: {
           closeOnClick: true,
           closeOnEsc: true,
           onBeforeLoad: function(event) {
             var overlay;
             overlay = this.getOverlay();
+            $("div.pb-ajax>div", overlay).addClass("container");
+            $("h3", overlay).remove();
+            $("textarea", overlay).remove();
+            $("input", overlay).remove();
             return overlay.draggable();
           },
           onLoad: function(event) {